+++ extensions/figma-files/raycast-env.d.ts
@@ -9,24 +9,24 @@
 
 type ExtensionPreferences = {
   /** Personal Access Token - Log in to Figma and click settings to setup a token */
-  "PERSONAL_ACCESS_TOKEN": string,
+  PERSONAL_ACCESS_TOKEN: string;
   /** Team ID/s - Find this in the url after navigating to your team page. Add multiple Team IDs separated by a comma. */
-  "TEAM_ID": string
-}
+  TEAM_ID: string;
+};
 
 /** Preferences accessible in all the extension's commands */
-declare type Preferences = ExtensionPreferences
+declare type Preferences = ExtensionPreferences;
 
 declare namespace Preferences {
   /** Preferences accessible in the `index` command */
-  export type Index = ExtensionPreferences & {}
+  export type Index = ExtensionPreferences & {};
   /** Preferences accessible in the `menu-bar` command */
-  export type MenuBar = ExtensionPreferences & {}
+  export type MenuBar = ExtensionPreferences & {};
 }
 
 declare namespace Arguments {
   /** Arguments passed to the `index` command */
-  export type Index = {}
+  export type Index = {};
   /** Arguments passed to the `menu-bar` command */
-  export type MenuBar = {}
+  export type MenuBar = {};
 }