+++ extensions/figma-files/src/hooks/useVisitedFiles.ts
@@ -25,8 +25,6 @@ export async function clearVisitedFiles() {
   return await LocalStorage.removeItem(VISITED_FIGMA_FILES_KEY);
 }
 
-
-
 export function useVisitedFiles() {
   const [files, setFiles] = useCachedState<File[]>("visitedFiles");